Install lizmap api as standard python package
