+++ .gitignore
@@ -10,3 +10,9 @@
 lizmap.cfg
 
 .idea/
+
+*.egg-info
+
+build/
+dist/
+
